Map Better
Added random event and unclaimed profiles. For random events and empty systems.

Maps now take from a pool. positions of nodes is still random, but the number of types of nodes is less random now.

replaced Nebula.random() with an init function that can take up to two arguments. Random if none, not random if one or two.