Map Gen Redo
wow. kinda rough. necessary though. means i dont have to create an actual end to the game, which is good. that would take a while.
The map now generates rank-by rank, meaning that it can be generated indefinitely.
Fixed a typo. Seperation?? what is that...
Map nodes are now generated at random, not through fixed pools.
Might make this approach the ratio of the weights in the future.
Reduced shimmy, increased separation on Y.
Prepared to save state of RNG in Global as to be able to access the map again.